demo-rpg-aiapp/dev/rpg-backend-python/requirements.txt
azure-functions
azure-identity
azure-keyvault-secrets
openai
pymssql
